+++ src/elements/GeoTree/GeoTree.constants.js
@@ -1,0 +1,9 @@
+const INPUT_MODES = {
+  SINGLE: 'single',
+  MULTIPLE: 'multiple',
+  HIDDEN: 'hidden'
+}
+
+export {
+  INPUT_MODES
+}